+++ packages/at_repl/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.1.0
+- feat: `--key-file|-k option to specify key file`
+- feat: `--version` to print at_repl version
+- feat: root domain options: `--root-domain`, `--rootDomain`, `--root-url`, 
+  `--rootUrl`, `-r`
+- feat: support proxy servers in root domain
+
 ## 2.0.0
 
 - feat: `/inspect` command